transferred WorkerAdmin and PositionAdmin into another app
